Created the abstractfactory
Created abstractfactory with HTML and MD file creator, along with
testing of it with the help of App.java. Also added type() method to
each Element-class so that the enum could be returned.


Some slight changes here and there to make everything work together with
the new expansion of the system.